User commit

+++ src/main/java/com/example/demo/service/RecipeIngredientService.java
@@ -1,9 +1,10 @@
 package com.example.demo.service;
 
 import com.example.demo.entity.RecipeIngredient;
-import java.math.BigDecimal;
 
 public interface RecipeIngredientService {
+
     RecipeIngredient addIngredientToMenuItem(RecipeIngredient recipeIngredient);
-    BigDecimal getTotalQuantityOfIngredient(Long ingredientId);
+
+    Double getTotalQuantityOfIngredient(Long ingredientId);
 }
